requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.16
django-allauth==65.5.0
django-ckeditor==5.9.0
django-cors-headers==4.7.0
django-debug-toolbar==3.7.0
django-environ==0.12.0
django-formtools==2.5.1
django-heroku==0.3.1
django-jazzmin==3.0.1
django-js-asset==1.2.2
django-otp==1.5.4
django-otp-yubikey==1.1.0
django-phonenumber-field==8.0.0
django-redis==5.2.0
django-taggit==1.2.0
django-two-factor-auth==1.17.0
django-user-visit==0.4
djangorestframework==3.15.2
frozenlist==1.5.0
gunicorn==20.0.4
idna==3.10
importlib-metadata==5.0.0
legacy-cgi==2.6.2
multidict==6.1.0
mysqlclient==2.2.7
packaging==21.3
phonenumbers==8.13.55
pillow==11.1.0
propcache==0.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycryptodome==3.22.0
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.0.9
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2019.3
qrcode==7.4.2
redis==4.3.4
requests==2.32.3
setuptools==76.1.0
sqlparse==0.3.1
twilio==9.5.0
typing_extensions==4.4.0
tzdata==2025.1
ua-parser==0.10.0
urllib3==2.3.0
user-agents==2.2.0
whitenoise==5.0.1
wrapt==1.14.1
yarl==1.18.3
YubiOTP==1.0.0
zipp==3.10.0
